LayoutTests/fast/canvas/webgl/largeBuffer-expected.txt

PASS gl.getError() is 0
PASS gl.getError() is 1281
PASS size is 0
PASS gl.getError() is 1281
PASS gl.getError() is 0
PASS successfullyParsed is true

TEST COMPLETE
